+++ ui/src/shared/components/TimeMachine/fluxQueryBuilder/TagSelector.tsx
@@ -28,6 +28,7 @@ import {
 import {ButtonShape, ComponentSize, IconFont} from 'src/reusable_ui/types'
 import {Button} from 'src/reusable_ui'
 import LoadingSpinner from 'src/reusable_ui/components/spinners/LoadingSpinner'
+import {hightlightDropdown} from '../../TimeRangeDropdown'
 
 const SEARCH_DEBOUNCE_MS = 400
 
@@ -39,7 +40,9 @@ function renderType(type: BuilderAggregateFunctionType) {
 }
 
 type Callbacks = ReturnType<typeof mdtp>
-type Props = TagSelectorState & Callbacks & TimeMachineQueryProps
+type Props = TagSelectorState &
+  Callbacks &
+  TimeMachineQueryProps & {timeRangeText: string}
 
 const TagSelector = (props: Props) => {
   const {
@@ -91,6 +94,7 @@ const TagSelectorBody = (props: Props) => {
     onIncreaseKeysLimit,
     onSearchKeys,
     tagValues: selectedValues,
+    timeRangeText,
   } = props
   if (aggregateFunctionType === 'filter') {
     if (keysStatus === RemoteDataState.Error) {
@@ -104,7 +108,13 @@ const TagSelectorBody = (props: Props) => {
     ) {
       return (
         <BuilderCard.Empty testID="empty-tag-keys">
-          No tag keys found <small>in the current time range</small>
+          No tag keys found{' '}
+          <small>
+            in the current{' '}
+            <a href="#" onClick={hightlightDropdown} title={timeRangeText}>
+              time range
+            </a>
+          </small>
         </BuilderCard.Empty>
       )
     }
@@ -205,6 +215,7 @@ const TagSelectorValues = (props: Props & {onLoadMoreValues: () => void}) => {
     tagValues: selectedValues,
     onSelectValues,
     onLoadMoreValues,
+    timeRangeText,
   } = props
   if (aggregateFunctionType === 'filter') {
     if (keysStatus === RemoteDataState.NotStarted) {
@@ -258,7 +269,13 @@ const TagSelectorValues = (props: Props & {onLoadMoreValues: () => void}) => {
     if (!values.length) {
       return (
         <BuilderCard.Empty>
-          No values found <small>in the current time range</small>
+          No values found{' '}
+          <small>
+            in the current{' '}
+            <a href="#" onClick={hightlightDropdown} title={timeRangeText}>
+              time range
+            </a>
+          </small>
         </BuilderCard.Empty>
       )
     }
@@ -289,7 +306,7 @@ const TagSelectorValues = (props: Props & {onLoadMoreValues: () => void}) => {
           )
         })}
         {valuesTruncated && aggregateFunctionType === 'filter' ? (
-          <div className="flux-query-builder--list-item" key={`__truncated`}>
+          <div className="flux-query-builder--list-item" key="__truncated">
             {valuesStatus !== RemoteDataState.Loading ? (
               <Button
                 text="Load More Values"
@@ -306,6 +323,18 @@ const TagSelectorValues = (props: Props & {onLoadMoreValues: () => void}) => {
             )}
           </div>
         ) : undefined}
+        <div
+          className="flux-query-builder--list-item"
+          key="__infomsg"
+          style={{cursor: 'default'}}
+        >
+          <small style={{textAlign: 'center', width: '100%'}}>
+            keys/values depend on{' '}
+            <a href="#" onClick={hightlightDropdown} title={timeRangeText}>
+              time range
+            </a>
+          </small>
+        </div>
       </div>
     </BuilderCard.Body>
   )